hr-bot/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.12.1

# LangChain & LangGraph
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langgraph==0.0.40

# 向量数据库
chromadb==0.4.18

# Embedding模型
sentence-transformers==2.2.2

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# 数据处理
pandas==2.1.3
openpyxl==3.1.2

# 日志
loguru==0.7.2

# 文档处理
pypdf==3.17.1
python-docx==1.1.0
